Add player skills

+++ Assets/Scripts/UI/ExperienceBar.cs
@@ -6,7 +6,7 @@ public class ExperienceBar : MonoBehaviour
     [SerializeField] RectTransform bar;
     [SerializeField] TextMeshProUGUI textComponent;
 
-    public void SetExperience(int level, int experience, int totalExp)
+    public void SetExperience(int level, float experience, float totalExp)
     {
         bar.localScale = new Vector3(experience / totalExp, 1, 1);
         textComponent.text = $"Lvl {level}";
